+++ ft_printf/ft_nbrconvert_len.c
@@ -6,15 +6,15 @@
 /*   By: pgiraude <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/19 20:04:32 by pgiraude          #+#    #+#             */
-/*   Updated: 2022/08/19 21:11:18 by pgiraude         ###   ########.fr       */
+/*   Updated: 2022/08/22 20:44:18 by pgiraude         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_nbrconvert_len(long int nbr, int base)
+size_t	ft_nbrconvert_len(size_t nbr, size_t base)
 {
-	unsigned long int	len;
+	size_t	len;
 
 	len = 1;
 	while (nbr >= base)
